+++ WheelWizard/Views/Components/WhWzLibrary/FriendsListItem.axaml
@@ -3,7 +3,7 @@
         xmlns:lang="clr-namespace:WheelWizard.Resources.Languages"
         xmlns:components="clr-namespace:WheelWizard.Views.Components"
         xmlns:miiVars="using:WheelWizard.MiiImages.Domain"
-        xmlns:behaviorComponent="clr-namespace:WheelWizard.Views.BehaviorComponent">
+        xmlns:behaviorComponent="clr-namespace:WheelWizard.Views.Patterns">
     <Design.PreviewWith>
         <StackPanel Spacing="6" Background="{StaticResource Neutral950}">
             <components:FriendsListItem Width="428" Height="124"
