+++ WheelWizard/Views/Components/WhWzLibrary/MiiBlock.axaml
@@ -2,7 +2,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:controls="using:WheelWizard.Views.Components"
         xmlns:miiVars="using:WheelWizard.MiiImages.Domain"
-        xmlns:behaviorComponent="clr-namespace:WheelWizard.Views.BehaviorComponent">
+        xmlns:behaviorComponent="clr-namespace:WheelWizard.Views.Patterns">
     <Design.PreviewWith>
         <StackPanel Spacing="13" Background="{StaticResource Neutral950}">
             <controls:MiiBlock Width="90" Height="90" IsGlobal="True"/>
