+++ WheelWizard/Views/Components/WhWzLibrary/MiiBlock.axaml.cs
@@ -5,7 +5,7 @@ using Avalonia.Input;
 using Avalonia.Interactivity;
 using WheelWizard.MiiImages;
 using WheelWizard.MiiImages.Domain;
-using WheelWizard.Views.BehaviorComponent;
+using WheelWizard.Views.Patterns;
 using WheelWizard.WiiManagement;
 using WheelWizard.WiiManagement.MiiManagement;
 using WheelWizard.WiiManagement.MiiManagement.Domain.Mii;
