+++ WheelWizard/Views/Components/WhWzLibrary/PlayerListItem.axaml
@@ -3,7 +3,7 @@
         xmlns:lang="clr-namespace:WheelWizard.Resources.Languages"
         xmlns:components="using:WheelWizard.Views.Components"
         xmlns:miiVars="using:WheelWizard.MiiImages.Domain"
-        xmlns:behaviorComponent="clr-namespace:WheelWizard.Views.BehaviorComponent">
+        xmlns:behaviorComponent="clr-namespace:WheelWizard.Views.Patterns">
     <Design.PreviewWith>
         <StackPanel Background="{StaticResource Neutral950}">
             <components:PlayerListItem Width="428" Height="124"
