+++ WheelWizard/Views/Pages/kitchenSink/KitchenSinkInputFieldsPage.axaml
@@ -2,7 +2,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:behaviorComp="clr-namespace:WheelWizard.Views.BehaviorComponent"
+             xmlns:behaviorComp="clr-namespace:WheelWizard.Views.Patterns"
              mc:Ignorable="d" d:DesignWidth="620" d:DesignHeight="520"
              x:Class="WheelWizard.Views.Pages.KitchenSink.KitchenSinkInputFieldsPage">
     <StackPanel Spacing="12">
