+++ WheelWizard/Views/Patterns/AspectGrid.axaml
@@ -3,7 +3,7 @@
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
       mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-      x:Class="WheelWizard.Views.BehaviorComponent.AspectGrid"
+      x:Class="WheelWizard.Views.Patterns.AspectGrid"
       SizeChanged="OnSizeChanged">
     <ContentPresenter x:Name="ContentHost" HorizontalAlignment="Center" VerticalAlignment="Center"/>
 </Grid>
