+++ WheelWizard/Views/Patterns/FeedbackTextBox.axaml
@@ -3,8 +3,8 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="400" d:DesignHeight="100"
-             x:Class="WheelWizard.Views.BehaviorComponent.FeedbackTextBox"
-             xmlns:behaviorComp="clr-namespace:WheelWizard.Views.BehaviorComponent"
+             x:Class="WheelWizard.Views.Patterns.FeedbackTextBox"
+             xmlns:behaviorComp="clr-namespace:WheelWizard.Views.Patterns"
              xmlns:components="clr-namespace:WheelWizard.Views.Components"
              x:DataType="behaviorComp:FeedbackTextBox"
              >
