+++ WheelWizard/Views/Patterns/MiiImages/MiiCarousel.axaml.cs
@@ -7,7 +7,7 @@ using WheelWizard.MiiImages;
 using WheelWizard.MiiImages.Domain;
 using WheelWizard.WiiManagement.MiiManagement.Domain.Mii;
 
-namespace WheelWizard.Views.BehaviorComponent;
+namespace WheelWizard.Views.Patterns;
 
 public partial class MiiCarousel : BaseMiiImage
 {
