+++ WheelWizard/Views/Patterns/VrHistoryGraph.axaml
@@ -2,12 +2,12 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:behaviorComp="clr-namespace:WheelWizard.Views.BehaviorComponent"
+             xmlns:behaviorComp="clr-namespace:WheelWizard.Views.Patterns"
              xmlns:components="clr-namespace:WheelWizard.Views.Components"
              mc:Ignorable="d"
              d:DesignWidth="560"
              d:DesignHeight="390"
-             x:Class="WheelWizard.Views.BehaviorComponent.VrHistoryGraph"
+             x:Class="WheelWizard.Views.Patterns.VrHistoryGraph"
              x:DataType="behaviorComp:VrHistoryGraph">
     <UserControl.Styles>
         <Style Selector="Border.profileStatCard">
