+++ WheelWizard/Views/Popups/PlayerProfileWindow.axaml
@@ -2,7 +2,7 @@
                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
                    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-                   xmlns:behavior="clr-namespace:WheelWizard.Views.BehaviorComponent"
+                   xmlns:behavior="clr-namespace:WheelWizard.Views.Patterns"
                    xmlns:components="clr-namespace:WheelWizard.Views.Components"
                    xmlns:lang="clr-namespace:WheelWizard.Resources.Languages"
                    xmlns:miiVars="using:WheelWizard.MiiImages.Domain"
